Move literals to _types.py and helper functions to _converters.py
