+++ python/src/pypalmsens/_data/data_value.py
@@ -5,8 +5,11 @@ from dataclasses import dataclass
 import PalmSens
 from typing_extensions import override
 
-from .._methods.types import cr_enum_to_string, pr_enum_to_string
-from ..settings import (
+from .._converters import (
+    cr_enum_to_string,
+    pr_enum_to_string,
+)
+from .._types import (
     AllowedCurrentRanges,
     AllowedPotentialRanges,
     AllowedReadingStatus,
